remove some boilerplate code from lagopus
