+++ src/global.mts
@@ -1,7 +1,5 @@
-import { LagopusElement, LagopusObjectBuffer } from "./primes.mjs";
+import { SolubleObjectData } from "./primes.mjs";
 import { Atom } from "./atom.mjs";
-import { createBuffer } from "./utils.mjs";
-import { rand } from "./math.mjs";
 
 export var atomDevice: Atom<GPUDevice> = new Atom(null);
 export var atomContext: Atom<GPUCanvasContext> = new Atom(null);
@@ -10,15 +8,7 @@ export var atomDepthTexture: Atom<GPUTexture> = new Atom(null);
 
 export var atomBufferNeedClear: Atom<boolean> = new Atom(true);
 
-export var atomLagopusTree: Atom<LagopusElement> = new Atom(null);
-
-// touch events
-
-export var atomMouseHoldingPaths = new Atom<number[][]>([]);
-
-export let atomObjectsTree = new Atom<LagopusElement>(null);
-
-export let atomObjectsBuffer = new Atom<LagopusObjectBuffer[]>([]);
+export var atomLagopusTree: Atom<SolubleObjectData> = new Atom(null);
 
 export function wLog<T extends any>(message: string, a: T): T {
   console.warn(message, a);
@@ -27,5 +17,4 @@ export function wLog<T extends any>(message: string, a: T): T {
 
 // prepare shared array called `base_points`
 
-export let atomPointsSize: Atom<number> = new Atom(0);
 export let atomPointsBuffer: Atom<GPUBuffer> = new Atom(null);
