+++ src/math.mts
@@ -1,4 +1,4 @@
-import { LagopusElement, V2, V3 } from "./primes.mjs";
+import { V2, V3 } from "./primes.mjs";
 
 export let square = (x: number): number => {
   return x * x;
