+++ src/primes.mts
@@ -1,13 +1,10 @@
-/** a simpler type for Dispatch */
-export type FnDispatch = (op: string, data: any) => void;
-
 /** 3D point */
 export type V3 = [number, number, number];
 
 /** 2D point */
 export type V2 = [number, number];
 
-export interface LagopusAttribute {
+export interface SolubleAttribute {
   field: string;
   format: GPUVertexFormat;
   size: number;
@@ -15,43 +12,12 @@ export interface LagopusAttribute {
   unitSize?: number;
 }
 
-export interface LagopusObjectOptions {
-  shader: string;
-  topology: GPUPrimitiveTopology;
-  attrsList: LagopusAttribute[];
-  data: Record<string, number[]>[];
-  hitRegion?: LagopusHitRegion;
-  indices?: number[];
-}
-
-export interface LagopusObjectData {
+export interface SolubleObjectData {
   type: "object";
   topology: GPUPrimitiveTopology;
   vertexBuffersDescriptors: Iterable<GPUVertexBufferLayout | null>;
   shaderModule: GPUShaderModule;
   vertexBuffers: GPUBuffer[];
   length: number;
-  hitRegion?: LagopusHitRegion;
   indices?: GPUBuffer;
-}
-
-export interface LagopusGroup {
-  type: "group";
-  children: LagopusElement[];
-}
-
-export type LagopusElement = LagopusGroup | LagopusObjectData;
-
-/** hitRegion information for object element */
-export interface LagopusHitRegion {
-  radius: number;
-  position: V3;
-  onHit?: (e: MouseEvent, d: (op: string, data: any) => void) => void;
-  onMousedown?: (e: MouseEvent, d: (op: string, data: any) => void) => void;
-  onMousemove?: (e: MouseEvent, d: (op: string, data: any) => void) => void;
-  onMouseup?: (e: MouseEvent, d: (op: string, data: any) => void) => void;
-}
-
-export interface LagopusObjectBuffer {
-  hitRegion?: LagopusHitRegion;
 }
